in work
)

+++ DevEduEducationSystem.API.Tests/Support/Models/ResponseModels/GroupResponseModel.cs
@@ -32,7 +32,8 @@ namespace DevEduEducationSystem.API.Tests.Support.Models
         public string Timetable { get; set; }
 
         [JsonPropertyName("paymentPerMonth")]
-        public int PaymentPerMonth { get; set; }
+        public decimal PaymentPerMonth { get; set; }
+
 
         public override bool Equals(object? obj)
         {
